+++ R/priors.R
@@ -707,15 +707,24 @@ prior_gp <- function(bterms, data, def_scale_prior) {
   #   def_scale_prior: a character string defining 
   #     the default prior for random effects SDs
   prior <- empty_brmsprior()
-  gpterms <- all_terms(bterms[["gp"]])
-  if (length(gpterms)) {
+  gpef <- tidy_gpef(bterms, data)
+  if (nrow(gpef)) {
     px <- check_prefix(bterms)
     lscale_prior <- def_lscale_prior(bterms, data)
+    gpterms <- gpef$term
+    gpterms_by <- named_list(gpterms, gpterms)
+    for (i in seq_along(gpterms_by)) {
+      # GPs of each 'by' level should get their own default lscale prior
+      if (length(gpef$bylevels[[i]])) {
+        str_add(gpterms_by[[i]]) <- paste0(":", gpef$bylevels[[i]])
+      }
+    }
+    gpterms_by <- unlist(gpterms_by)
     prior <- prior +
       brmsprior(class = "sdgp", prior = def_scale_prior, ls = px) +
       brmsprior(class = "sdgp", coef = gpterms, ls = px) +
       brmsprior(class = "lscale", prior = "normal(0, 0.5)", ls = px) +
-      brmsprior(class = "lscale", prior = lscale_prior, coef = gpterms, ls = px)
+      brmsprior(class = "lscale", prior = lscale_prior, coef = gpterms_by, ls = px)
   }
   prior
 }