+++ NEWS.md
@@ -4,6 +4,9 @@
 
 * Define custom variables in all of Stan's program blocks
 via function `stanvar`. (#459)
+* Allow to automatically group predictor values in Gaussian 
+processes specified via `gp`. This may lead to a
+considerable increase in sampling efficiency. (#300)
 * Compute LOO-adjusted R-squared using method `loo_R2`.
 * Allow computing non-linear predictors outside of a loop over 
 observations by means of argument `loop` in `brmsformula`.